Split the definitions for each glyph type into a separate file
