+++ datashader/glyphs/__init__.py
@@ -1,0 +1,26 @@
+from __future__ import absolute_import
+from .points import Point
+from .line import (
+    LineAxis0,
+    LineAxis0Multi,
+    LinesAxis1,
+    LinesAxis1XConstant,
+    LinesAxis1YConstant,
+    LinesAxis1Ragged,
+)
+from .area import (
+    AreaToZeroAxis0,
+    AreaToZeroAxis0Multi,
+    AreaToZeroAxis1,
+    AreaToZeroAxis1XConstant,
+    AreaToZeroAxis1YConstant,
+    AreaToZeroAxis1Ragged,
+    AreaToLineAxis0,
+    AreaToLineAxis0Multi,
+    AreaToLineAxis1,
+    AreaToLineAxis1XConstant,
+    AreaToLineAxis1YConstant,
+    AreaToLineAxis1Ragged,
+)
+from .trimesh import Triangles
+from .glyph import Glyph
